+++ mayan/apps/document_signatures/tests/test_views.py
@@ -2,11 +2,8 @@ from __future__ import absolute_import, unicode_literals
 
 import logging
 
-from django.core.files import File
-
 from django_downloadview.test import assert_download_response
 
-from mayan.apps.django_gpg.models import Key
 from mayan.apps.documents.models import DocumentVersion
 from mayan.apps.documents.tests import (
     GenericDocumentViewTestCase, TEST_DOCUMENT_PATH
@@ -21,15 +18,14 @@ from ..permissions import (
     permission_document_version_signature_view
 )
 
-from .literals import (
-    TEST_SIGNATURE_FILE_PATH, TEST_SIGNED_DOCUMENT_PATH, TEST_KEY_FILE
-)
+from .literals import TEST_SIGNATURE_FILE_PATH, TEST_SIGNED_DOCUMENT_PATH
+from .mixins import SignaturesTestMixin
 
 TEST_UNSIGNED_DOCUMENT_COUNT = 4
 TEST_SIGNED_DOCUMENT_COUNT = 2
 
 
-class SignaturesViewTestCase(GenericDocumentViewTestCase):
+class SignaturesViewTestCase(SignaturesTestMixin, GenericDocumentViewTestCase):
     def setUp(self):
         super(SignaturesViewTestCase, self).setUp()
         self.login_user()
@@ -41,251 +37,198 @@ class SignaturesViewTestCase(GenericDocumentViewTestCase):
         )
 
     def test_signature_list_view_no_permission(self):
-        with open(TEST_KEY_FILE, mode='rb') as file_object:
-            Key.objects.create(key_data=file_object.read())
+        self._create_test_key()
 
         with open(TEST_DOCUMENT_PATH, mode='rb') as file_object:
-            document = self.document_type.new_document(
+            self.document = self.document_type.new_document(
                 file_object=file_object
             )
 
-        with open(TEST_SIGNATURE_FILE_PATH, mode='rb') as file_object:
-            DetachedSignature.objects.create(
-                document_version=document.latest_version,
-                signature_file=File(file_object)
-            )
+        self._create_detached_signature()
 
         response = self._request_document_version_signature_list_view(
-            document=document
+            document=self.document
         )
         self.assertEqual(response.status_code, 403)
 
     def test_signature_list_view_with_access(self):
-        with open(TEST_KEY_FILE, mode='rb') as file_object:
-            Key.objects.create(key_data=file_object.read())
+        self._create_test_key()
 
         with open(TEST_DOCUMENT_PATH, mode='rb') as file_object:
-            document = self.document_type.new_document(
+            self.document = self.document_type.new_document(
                 file_object=file_object
             )
 
-        with open(TEST_SIGNATURE_FILE_PATH, mode='rb') as file_object:
-            DetachedSignature.objects.create(
-                document_version=document.latest_version,
-                signature_file=File(file_object)
-            )
+        self._create_detached_signature()
 
         self.grant_access(
-            obj=document,
+            obj=self.document,
             permission=permission_document_version_signature_view
         )
 
         response = self._request_document_version_signature_list_view(
-            document=document
+            document=self.document
         )
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.context['object_list'].count(), 1)
 
-    def _request_document_version_signature_details_view(self, signature):
+    def _request_document_version_signature_details_view(self):
         return self.get(
             viewname='signatures:document_version_signature_details',
-            args=(signature.pk,)
+            args=(self.test_signature.pk,)
         )
 
     def test_signature_detail_view_no_permission(self):
-        with open(TEST_KEY_FILE, mode='rb') as file_object:
-            Key.objects.create(key_data=file_object.read())
+        self._create_test_key()
 
         with open(TEST_DOCUMENT_PATH, mode='rb') as file_object:
-            document = self.document_type.new_document(
+            self.document = self.document_type.new_document(
                 file_object=file_object
             )
 
-        with open(TEST_SIGNATURE_FILE_PATH, mode='rb') as file_object:
-            signature = DetachedSignature.objects.create(
-                document_version=document.latest_version,
-                signature_file=File(file_object)
-            )
+        self._create_detached_signature()
 
-        response = self._request_document_version_signature_details_view(
-            signature=signature
-        )
+        response = self._request_document_version_signature_details_view()
         self.assertEqual(response.status_code, 403)
 
     def test_signature_detail_view_with_access(self):
-        with open(TEST_KEY_FILE, mode='rb') as file_object:
-            Key.objects.create(key_data=file_object.read())
+        self._create_test_key()
 
         with open(TEST_DOCUMENT_PATH, mode='rb') as file_object:
-            document = self.document_type.new_document(
+            self.document = self.document_type.new_document(
                 file_object=file_object
             )
 
-        with open(TEST_SIGNATURE_FILE_PATH, mode='rb') as file_object:
-            signature = DetachedSignature.objects.create(
-                document_version=document.latest_version,
-                signature_file=File(file_object)
-            )
+        self._create_detached_signature()
 
         self.grant_access(
-            obj=document,
+            obj=self.document,
             permission=permission_document_version_signature_view
         )
 
-        response = self._request_document_version_signature_details_view(
-            signature=signature
-        )
+        response = self._request_document_version_signature_details_view()
         self.assertContains(
-            response=response, text=signature.signature_id, status_code=200
+            response=response, text=self.test_signature.signature_id,
+            status_code=200
         )
 
-    def _request_document_version_signature_upload_view(self, document_version):
+    def _request_document_version_signature_upload_view(self):
         with open(TEST_SIGNATURE_FILE_PATH, mode='rb') as file_object:
             return self.post(
                 viewname='signatures:document_version_signature_upload',
-                args=(document_version.pk,),
+                kwargs={'pk': self.document.latest_version.pk},
                 data={'signature_file': file_object}
             )
 
     def test_signature_upload_view_no_permission(self):
         with open(TEST_DOCUMENT_PATH, mode='rb') as file_object:
-            document = self.document_type.new_document(
+            self.document = self.document_type.new_document(
                 file_object=file_object
             )
 
-        response = self._request_document_version_signature_upload_view(
-            document_version=document.latest_version
-        )
+        response = self._request_document_version_signature_upload_view()
         self.assertEqual(response.status_code, 403)
         self.assertEqual(DetachedSignature.objects.count(), 0)
 
     def test_signature_upload_view_with_access(self):
         with open(TEST_DOCUMENT_PATH, mode='rb') as file_object:
-            document = self.document_type.new_document(
+            self.document = self.document_type.new_document(
                 file_object=file_object
             )
 
         self.grant_access(
-            obj=document,
+            obj=self.document,
             permission=permission_document_version_signature_upload
         )
 
-        response = self._request_document_version_signature_upload_view(
-            document_version=document.latest_version
-        )
+        response = self._request_document_version_signature_upload_view()
         self.assertEqual(response.status_code, 302)
         self.assertEqual(DetachedSignature.objects.count(), 1)
 
-    def _request_document_version_signature_download_view(self, signature):
+    def _request_document_version_signature_download_view(self):
         return self.get(
             viewname='signatures:document_version_signature_download',
-            args=(signature.pk,),
+            kwargs={'pk': self.test_signature.pk}
         )
 
     def test_signature_download_view_no_permission(self):
         with open(TEST_DOCUMENT_PATH, mode='rb') as file_object:
-            document = self.document_type.new_document(
+            self.document = self.document_type.new_document(
                 file_object=file_object
             )
 
-        with open(TEST_SIGNATURE_FILE_PATH, mode='rb') as file_object:
-            signature = DetachedSignature.objects.create(
-                document_version=document.latest_version,
-                signature_file=File(file_object)
-            )
+        self._create_detached_signature()
 
-        response = self._request_document_version_signature_download_view(
-            signature=signature
-        )
+        response = self._request_document_version_signature_download_view()
         self.assertEqual(response.status_code, 403)
 
     def test_signature_download_view_with_access(self):
         with open(TEST_DOCUMENT_PATH, mode='rb') as file_object:
-            document = self.document_type.new_document(
+            self.document = self.document_type.new_document(
                 file_object=file_object
             )
 
-        with open(TEST_SIGNATURE_FILE_PATH, mode='rb') as file_object:
-            signature = DetachedSignature.objects.create(
-                document_version=document.latest_version,
-                signature_file=File(file_object)
-            )
+        self._create_detached_signature()
 
         self.grant_access(
-            obj=document,
+            obj=self.document,
             permission=permission_document_version_signature_download
         )
 
         self.expected_content_type = 'application/octet-stream; charset=utf-8'
 
-        response = self._request_document_version_signature_download_view(
-            signature=signature
-        )
+        response = self._request_document_version_signature_download_view()
 
-        with signature.signature_file as file_object:
+        with self.test_signature.signature_file as file_object:
             assert_download_response(
                 self, response=response, content=file_object.read(),
             )
 
-    def _request_document_version_signature_delete_view(self, signature):
+    def _request_document_version_signature_delete_view(self):
         return self.post(
             viewname='signatures:document_version_signature_delete',
-            args=(signature.pk,)
+            kwargs={'pk': self.test_signature.pk}
         )
 
     def test_signature_delete_view_no_permission(self):
-        with open(TEST_KEY_FILE, mode='rb') as file_object:
-            Key.objects.create(key_data=file_object.read())
+        self._create_test_key()
 
         with open(TEST_DOCUMENT_PATH, mode='rb') as file_object:
-            document = self.document_type.new_document(
+            self.document = self.document_type.new_document(
                 file_object=file_object
             )
 
-        with open(TEST_SIGNATURE_FILE_PATH, mode='rb') as file_object:
-            signature = DetachedSignature.objects.create(
-                document_version=document.latest_version,
-                signature_file=File(file_object)
-            )
+        self._create_detached_signature()
 
         self.grant_access(
-            obj=document,
+            obj=self.document,
             permission=permission_document_version_signature_view
         )
 
-        response = self._request_document_version_signature_delete_view(
-            signature=signature
-        )
+        response = self._request_document_version_signature_delete_view()
         self.assertEqual(response.status_code, 403)
         self.assertEqual(DetachedSignature.objects.count(), 1)
 
     def test_signature_delete_view_with_access(self):
-        with open(TEST_KEY_FILE, mode='rb') as file_object:
-            Key.objects.create(key_data=file_object.read())
+        self._create_test_key()
 
         with open(TEST_DOCUMENT_PATH, mode='rb') as file_object:
-            document = self.document_type.new_document(
+            self.document = self.document_type.new_document(
                 file_object=file_object
             )
 
-        with open(TEST_SIGNATURE_FILE_PATH, mode='rb') as file_object:
-            signature = DetachedSignature.objects.create(
-                document_version=document.latest_version,
-                signature_file=File(file_object)
-            )
+        self._create_detached_signature()
 
         self.grant_access(
-            obj=document,
+            obj=self.document,
             permission=permission_document_version_signature_delete
         )
         self.grant_access(
-            obj=document,
+            obj=self.document,
             permission=permission_document_version_signature_view
         )
 
-        response = self._request_document_version_signature_delete_view(
-            signature=signature
-        )
+        response = self._request_document_version_signature_delete_view()
         self.assertEqual(response.status_code, 302)
         self.assertEqual(DetachedSignature.objects.count(), 0)